+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.comd/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.orgd/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- allow horizontal layout by setting `radioGroup.isVertical = false`
+
 ## [1.2.5] - 2019-07-13
 
 ### Changed